requirements.txt
django==3.0
djangorestframework==3.12.2
psycopg2-binary==2.8.6
pytest==7.1.3
pytest-django==4.5.2
factory-boy==3.2.1
requests==2.21.0
black==22.8.0